Move all keyboard shortcut handling to `CourseShell`
